requirements.txt
# Backend dependencies for Databricks Genie Integration App
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
databricks-sdk>=0.59.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.1.1